% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_crs.R
\name{convert_crs}
\alias{convert_crs}
\title{Conversion to a New Coordinate Reference System}
\usage{
convert_crs(LatList, LongList, InitProj, NextProj = NULL)
}
\arguments{
\item{LatList}{A vector of size \emph{n} for latitudes collected from a GPS recording device}

\item{LongList}{A vector of size \emph{n} for longitudes collected from a GPS recording device}

\item{InitProj}{A character string of initial projection arguments; the arguments must be entered exactly as in the \href{https://proj4.org/usage/projections.html}{PROJ.4 documentation}}

\item{NextProj}{(Optional) A character string of final projection arguments; the arguments must be entered exactly as in the \href{https://proj4.org/usage/projections.html}{PROJ.4 documentation} (if \code{NULL}, the \href{https://www.openstreetmap.org}{OpenStreetMap} projection will be assumed)}
}
\value{
\code{\link{convert_crs}} return a list of 2 vectors of size \emph{n} for the converted latitude and longitude
}
\description{
Convert the coordinates of the recorded GPS data in an initial CRS to any desired CRS
}
\examples{

LatList <- c(45.86880,45.86887,45.86924,
45.87158,45.87014,45.86923,45.86808,
45.87177,45.87020,45.88010)

LongList <- c(7.173500,7.172540,7.171636,
7.180180,7.178070,7.177229,7.175240,
7.181409,7.179299,7.172235)

InitProj <- "+init=epsg:4326"

#converts the GPS coordinates to OSM CRS
convert_crs(LatList,LongList,InitProj)

NextProj <- "+init=epsg:4121 +proj=longlat +ellps=GRS80 +datum=GGRS87 +no_defs +towgs84=-199.87,74.79,246.62"

#converts the GPS coordinates to a selected projection
convert_crs(LatList,LongList,InitProj,NextProj)

}
