% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_coords.R
\name{interpolate_coords}
\alias{interpolate_coords}
\title{Coordinates Interpolation}
\usage{
interpolate_coords(LatList, LongList, timeseq = NULL, timeint = 1)
}
\arguments{
\item{LatList}{list of latitudes collected from a GPS recording device}

\item{LongList}{list of longitudes collected from a GPS recording device}

\item{timeseq}{list of irregular time series for GPS recording device in
format \code{"2000-01-01 00:00:00"} (default is \code{NULL} for no available
sequence of recording time)}

\item{timeint}{desired time step or increment of the time sequence in seconds
(default value is \code{1} second)}
}
\value{
\code{interpolate_coords} return a matrix of interpolated regularized
time series, interpolated latitude, and interpolated longitude.
}
\description{
Generate list of interpolated latitudes and longitudes by regularization of the
time series for a specified increment \code{timeint}
}
\examples{
LatList <- c(31.67514,31.675195,31.67525,31.675304,31.675356,31.675408,31.675467,
31.675517,31.675569,31.675623)
LongList <- c(-106.326522,-106.326367,-106.326211,-106.326058,-106.325901,
-106.325739,-106.325572,-106.32541,-106.325247,-106.325092)
timeseq <- c("2019-04-29 15:20:51","2019-04-29 15:21:03","2019-04-29 15:21:06",
"2019-04-29 15:21:15","2019-04-29 15:21:17","2019-04-29 15:21:32","2019-04-29 15:21:34",
"2019-04-29 15:21:51","2019-04-29 15:22:09","2019-04-29 15:22:36")

interpolate_coords(LatList,LongList,timeseq,1)   #regularize time sequence over 1 second
interpolate_coords(LatList,LongList,timeseq,60)   #regularize time sequence over 1 minute
}
