% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linklocate.R
\name{linklocate}
\alias{linklocate}
\title{Title}
\usage{
linklocate(LatList, LongList)
}
\arguments{
\item{LatList}{-List of latitudes collected from a GPS recording device}

\item{LongList}{-List of longitudes collected from a GPS recording device}
}
\value{
A list of correct latitutudes and longitudes mirrored on the corresponding link
}
\description{
Title
}
\examples{
LatList <- c(31.67514,31.675195,31.67525,31.675304,31.675356,31.675408,31.675467,31.675517,31.675569,31.675623)
LongList <- c(-106.326522,-106.326367,-106.326211,-106.326058,-106.325901,-106.325739,-106.325572,-106.32541,-106.325247,-106.325092)
linklocate(LatList,LongList)

}
