% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_highway.R
\name{match_highway}
\alias{match_highway}
\title{Coordinates Map Matching}
\usage{
match_highway(LatList, LongList, timeseq, k, boxcuts = get_boxes(LatList,
  LongList, timeseq, resolution, offLong, offLat)$boxtable$boxcuts,
  boxlist = get_boxes(LatList, LongList, timeseq, resolution, offLong,
  offLat)$boxlist, resolution = 10, offLong = 0.001, offLat = 0.001,
  osmlink = "https://api.openstreetmap.org/api/0.6/")
}
\arguments{
\item{LatList}{list of latitudes collected from a GPS recording device}

\item{LongList}{list of longitudes collected from a GPS recording device}

\item{timeseq}{list of time series for GPS recording device in format \code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{k}{maximum number of close highways to consider for the KNN analysis of each GPS coordinate}

\item{boxcuts}{a list of bounding box group number for each latitude and longitude (should be the same length as \code{LatList} and \code{LatList}; it can be generated from the \code{\link{get_boxes}} function by \code{~boxtable$boxcuts})}

\item{boxlist}{a list of bounding boxes coordinations; each bounding box coordinates is in the form of \code{("left","bottom","right","top")} (it can be generated from the \code{\link{get_boxes}} function by \code{~boxlist})}

\item{resolution}{an approximation of GPS recorded distance within each desired bounding box in kilometers (to be used for \code{\link{get_boxes}})}

\item{offLong}{a positive bounding box longitudal margin in decimal degrees (to be used for \code{\link{get_boxes}})}

\item{offLat}{a positive bounding box latitudal margin in decimal degrees (to be used for \code{\link{get_boxes}})}

\item{osmlink}{the API link for the \href{https://www.openstreetmap.org}{OpenStreetMap} data}
}
\value{
\code{\link{match_highway}} return a list of highway link IDs based on \href{https://www.openstreetmap.org}{OpenStreetMap} specification for the given GPS coordinates
}
\description{
Snap a set of GPS coordinates to the \href{https://www.openstreetmap.org}{OpenStreetMap (OSM)} data highways. This function facilitates the common process in two ways:
\itemize{
  \item Dividing the whole region into bounding boxes, which allows using the OSM data through API sourcing instead of setting up an OSM data server
  \item implement the fast K-Nearest Neighbor method to find the closest K links to each GPS coordinate
}
}
\examples{

LatList <- c(31.67514,31.675195,31.67525,
31.675304,31.675356,31.675408,31.675467,
31.675517,31.675569,31.675623)

LongList <- c(-106.326522,-106.326367,
-106.326211,-106.326058,-106.325901,
-106.325739,-106.325572,-106.32541,
-106.325247,-106.325092)

timeseq <- c("2019-04-29 15:20:51",
"2019-04-29 15:21:03","2019-04-29 15:21:06",
"2019-04-29 15:21:15","2019-04-29 15:21:17",
"2019-04-29 15:21:32","2019-04-29 15:21:34",
"2019-04-29 15:21:51","2019-04-29 15:22:09",
"2019-04-29 15:22:36")

timeseq <- as.POSIXct(timeseq)

k=5

resolution <- 0.1
offLong=0.001
offLat=0.002

match_highway(LatList, LongList, timeseq, k,
resolution = resolution,offLong = offLong, offLat = offLat)

boxcuts <- c(1,1,1,2,2,3,3,3,4,4)
boxlist <- cbind(c(-106.32752,31.67414,-106.28746,31.69790),
c(-106.26872,31.73050,-106.26627,31.77233),
c(-106.26915,31.77057,-106.26671,31.81236),
c(-106.36679,31.81582,-106.34707,31.84163))

match_highway(LatList, LongList, timeseq,
k, boxcuts=boxcuts, boxlist=boxlist)
}
