% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route.R
\name{plot_route}
\alias{plot_route}
\title{GPS Route Plot}
\usage{
plot_route(LatList, LongList, timeseq,
  boxlist = as.matrix(osmar::corner_bbox(min(LongList), min(LatList),
  max(LongList), max(LatList))))
}
\arguments{
\item{LatList}{A vector of size \emph{n} for latitudes collected from a GPS recording device}

\item{LongList}{A vector of size \emph{n} for longitudes collected from a GPS recording device}

\item{timeseq}{A vector of size \emph{n} for irregular time sequence of recorded GPS data in format \code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{boxlist}{(Optional) A matrix of size \emph{4 X p} of bounding boxes coordinates with rows of \code{("left","bottom","right","top")}; it can be generated from the \code{\link{get_boxes}} function by \code{~boxlist})}
}
\value{
map of the route including the average speed contour and bounding boxes
}
\description{
GPS Route Plot
}
\examples{

LatList <- c(31.67530, 31.67541, 31.67653,
31.67674, 31.67692, 31.67781, 31.67827,
31.67865, 31.68031, 31.68185, 31.68241,
31.68338, 31.68434, 31.68495, 31.68702,
31.68853, 31.68910, 31.68967, 31.69316, 31.69408)

LongList <- c(-106.3261, -106.3257, -106.3218,
-106.3210, -106.3204, -106.3174, -106.3164,
-106.3156, -106.3123, -106.3092, -106.3082,
-106.3059, -106.3036, -106.3022, -106.2978,
-106.2962, -106.2957, -106.2951, -106.2919,
-106.2911)

timeseq <- c("2019-04-29 15:09:59",
"2019-04-29 15:10:01","2019-04-29 15:10:23",
"2019-04-29 15:10:27","2019-04-29 15:10:30",
"2019-04-29 15:10:44","2019-04-29 15:10:49",
"2019-04-29 15:10:53","2019-04-29 15:11:10",
"2019-04-29 15:11:26","2019-04-29 15:11:32",
"2019-04-29 15:11:43","2019-04-29 15:11:53",
"2019-04-29 15:11:59","2019-04-29 15:12:19",
"2019-04-29 15:12:28","2019-04-29 15:12:31",
"2019-04-29 15:12:34","2019-04-29 15:12:53",
"2019-04-29 15:12:58")

timeseq <- as.POSIXct(timeseq)

plot_route(LatList,LongList,timeseq)

InterpolatedData <- interpolate_coords(LatList,LongList,timeseq,1)
LatList2 <- InterpolatedData$Latitude
LongList2 <- InterpolatedData$Longitude
timeseq2 <- InterpolatedData$DateTime

plot_route(LatList2,LongList2,timeseq2)

}
