% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_highway.R
\name{match_highway}
\alias{match_highway}
\title{Coordinates Map Matching}
\usage{
match_highway(LatList, LongList, timeseq, k,
  boxcuts = as.numeric(get_boxes(LatList, LongList, timeseq, resolution,
  offLong, offLat)$boxtable$boxcuts),
  boxlist = as.matrix(get_boxes(LatList, LongList, timeseq, resolution,
  offLong, offLat)$boxlist), resolution = 10, offLong = 0.001,
  offLat = 0.001, osmlink = "https://api.openstreetmap.org/api/0.6/")
}
\arguments{
\item{LatList}{A vector of size \emph{n} for latitudes collected from a GPS recording device}

\item{LongList}{A vector of size \emph{n} for longitudes collected from a GPS recording device}

\item{timeseq}{A vector of size \emph{n} for irregular time sequence of recorded GPS data in format \code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{k}{Value of maximum number of close highways; to consider for the KNN analysis}

\item{boxcuts}{(Optional) A vector of size \emph{n} with \emph{p} levels for the bounding box split of GPS coordinates; it can be generated from the \code{\link{get_boxes}} function by \code{~boxtable$boxcuts})}

\item{boxlist}{(Optional) A matrix of size \emph{4 X p} of bounding boxes coordinates with rows of \code{("left","bottom","right","top")}; it can be generated from the \code{\link{get_boxes}} function by \code{~boxlist})}

\item{resolution}{(Optional) Approximate value of distance within each bounding box in kilometers; to be used for \code{\link{get_boxes}}}

\item{offLong}{(Optional) A positive value of bounding box longitudal margin in decimal degrees; to be used for \code{\link{get_boxes}}}

\item{offLat}{(Optional) A positive value of bounding box latitudal margin in decimal degrees; to be used for \code{\link{get_boxes}}}

\item{osmlink}{(Optional) The API link for the \href{https://www.openstreetmap.org}{OpenStreetMap} data}
}
\value{
\code{\link{match_highway}} return a vector of size \emph{n} for highway link IDs based on \href{https://www.openstreetmap.org}{OpenStreetMap}
}
\description{
Snap a set of GPS coordinates in WGS84 to the \href{https://www.openstreetmap.org}{OpenStreetMap (OSM)} data highways. This function facilitates the common process in two ways:
\itemize{
  \item Dividing the whole region into bounding boxes, which allows using the OSM data through API sourcing instead of setting up an OSM data server
  \item Implementation of a fast K-Nearest Neighbor method to find the closest K links to each GPS coordinate
}
}
\examples{

LatList <- c(31.67514,31.675195,31.67525,
31.675304,31.675356,31.675408,31.675467,
31.675517,31.675569,31.675623)

LongList <- c(-106.326522,-106.326367,
-106.326211,-106.326058,-106.325901,
-106.325739,-106.325572,-106.32541,
-106.325247,-106.325092)

timeseq <- c("2019-04-29 15:20:51",
"2019-04-29 15:21:03","2019-04-29 15:21:06",
"2019-04-29 15:21:15","2019-04-29 15:21:17",
"2019-04-29 15:21:32","2019-04-29 15:21:34",
"2019-04-29 15:21:51","2019-04-29 15:22:09",
"2019-04-29 15:22:36")

timeseq <- as.POSIXct(timeseq)

k=5

resolution <- 0.1
offLong=0.001
offLat=0.002

match_highway(LatList, LongList, timeseq, k,
resolution = resolution,offLong = offLong, offLat = offLat)

boxcuts <- c(1,1,1,2,2,3,3,3,4,4)
boxlist <- cbind(c(-106.32752,31.67414,-106.28746,31.69790),
c(-106.26872,31.73050,-106.26627,31.77233),
c(-106.26915,31.77057,-106.26671,31.81236),
c(-106.36679,31.81582,-106.34707,31.84163))

match_highway(LatList, LongList, timeseq,
k, boxcuts=boxcuts, boxlist=boxlist)
}
